Fix faulty type

+++ modules/base/rendering/renderablenodedirectionhint.cpp
@@ -444,8 +444,8 @@ void RenderableNodeDirectionHint::updateShapeTransforms() {
     // TODO: check that distance is larger than arrow length?
 
     glm::dvec3 arrowDirection = glm::normalize(endNodePos - startNodePos);
-    glm::dvec3 startPos = glm::vec3(startNodePos + offset * arrowDirection);
-    glm::dvec3 endPos = glm::vec3(startPos + length * arrowDirection);
+    glm::dvec3 startPos = glm::dvec3(startNodePos + offset * arrowDirection);
+    glm::dvec3 endPos = glm::dvec3(startPos + length * arrowDirection);
 
     if (_invertArrowDirection) {
         std::swap(startPos, endPos);
